Move tests + use nofollow
